Extract test service setup into helper
