+++ test/test-meshctl-status.js
@@ -1,12 +1,11 @@
-var async = require('async');
+var assert = require('assert');
 var exec = require('./exec-meshctl');
 var test = require('tap').test;
+var testCmdHelper = require('./meshctl-helper');
 var util = require('util');
-var meshServer = require('../index').meshServer;
 var ServiceManager = require('../index').ServiceManager;
-var Client = require('../index').Client;
 
-test('Test actions', function(t) {
+test('Test status command', function(t) {
   function TestServiceManager() {
   }
   util.inherits(TestServiceManager, ServiceManager);
@@ -52,58 +51,14 @@ test('Test actions', function(t) {
     ''
   ].join('\n');
 
-  var server = meshServer(new TestServiceManager());
-  server.set('port', 0);
-  server.start(function(err, port) {
-    t.ifError(err);
-
-    t.test('Setup base models', function(tt) {
-      var service = {
-        name: 'service 1',
-        _groups: [{id: 1, name: 'group 1', scale: 1}]
-      };
-      var exec = {
-        address: '127.0.0.1',
-        APIPort: 5000,
-        totalCapacity: 2,
-      };
-      var inst = {
-        serverServiceId: 1,
-        groupId: 1,
-        executorId: 1,
-      };
-
-      var ServerService = server.models.ServerService;
-      var ServiceInstance = server.models.ServiceInstance;
-      var Executor = server.models.Executor;
-      async.series([
-        ServerService.create.bind(ServerService, service),
-        Executor.create.bind(Executor, exec),
-        ServiceInstance.create.bind(ServiceInstance, inst),
-      ], function(err) {
-        tt.ifError(err);
-        tt.end();
-      });
-    });
-
-    var instance = null;
-
-    t.test('Initialize client models', function(tt) {
-      function ctlRequest(service, instance, req, callback) {
-        tt.deepEqual(req, {cmd: 'status'}, 'Request should match');
+  testCmdHelper(t, TestServiceManager, function(t, service, instance, port) {
+    t.test('Setup service manager', function(tt) {
+      function ctlRequest(s, i, req, callback) {
+        assert.deepEqual(req, {cmd: 'status'}, 'Request should match');
         callback(null, STATUS);
       }
       TestServiceManager.prototype.ctlRequest = ctlRequest;
-
-      var client = new Client('http://127.0.0.1:' + port);
-      client.serviceFind('1', function(err, s) {
-        tt.ifError(err, 'Default service should be found');
-        s.instances.findById('1', function(err, i) {
-          tt.ifError(err, 'Default instance should be found');
-          instance = i;
-          tt.end();
-        });
-      });
+      tt.end();
     });
 
     t.test('Status summary API', function(tt) {
@@ -117,16 +72,10 @@ test('Test actions', function(t) {
     t.test('Status summary CLI', function(tt) {
       exec.resetHome();
       exec(port, 'status', function(err, stdout) {
-        tt.ifError(err, 'command shoudl not error');
+        tt.ifError(err, 'command should not error');
         tt.equal(RENDERED_STATUS, stdout, 'Rendered status should match');
         tt.end();
       });
     });
-
-    t.test('Cleanup', function(tt) {
-      server.stop(tt.end.bind(tt));
-    });
-
-    t.end();
   });
 });
